- Fix moving Parts from Step to Step so that if we leave nothing except a Submodel behind, we reset the PLI properly. - Fix swapping Steps so we correctly reorder Submodels. - Fix moving Steps from Page to Page so we correctly reorder Submodels. - Fix crash in new backup save setup, so we correctly handle non-existent 'File->Save As' file.

+++ src/Lic.py
@@ -1250,7 +1250,8 @@ class LicWindow(QMainWindow):
 
             if os.path.isfile(tmpName):
                 os.remove(tmpName)
-            os.rename(self.filename, tmpName)
+            if os.path.isfile(self.filename):
+                os.rename(self.filename, tmpName)
             os.rename(tmpXName, self.filename)
 
             self.undoStack.setClean()
